re-organize ef reparams
