+++ cirkit/new/reparams/__init__.py
@@ -1,12 +1,12 @@
 from .binary import BinaryReparam as BinaryReparam
-from .binary import EFProductReparam as EFProductReparam
 from .binary import KroneckerReparam as KroneckerReparam
 from .composed import ComposedReparam as ComposedReparam
+from .ef import EFNormalReparam as EFNormalReparam
+from .ef import EFProductReparam as EFProductReparam
 from .leaf import LeafReparam as LeafReparam
 from .normalized import LogSoftmaxReparam as LogSoftmaxReparam
 from .normalized import SoftmaxReparam as SoftmaxReparam
 from .reparam import Reparameterization as Reparameterization
-from .special import EFNormalReparam as EFNormalReparam
 from .unary import ClampReparam as ClampReparam
 from .unary import ExpReparam as ExpReparam
 from .unary import LinearReparam as LinearReparam
